Moved files to correct folders
